Localization

+++ frontend/components/Post/Post.js
@@ -8,6 +8,7 @@ import { postComment } from "@/redux/action/postsAction";
 import moment from "moment";
 import { useForm } from "react-hook-form";
 import EditModal from "../Modal/EditModal";
+import useTranslation from "next-translate/useTranslation";
 
 function Post(props) {
   const [show, setShow] = useState(false);
@@ -25,6 +26,7 @@ function Post(props) {
     handleSubmit,
     formState: { errors },
   } = useForm();
+  const { t } = useTranslation();
 
   useEffect(() => {
     if (!router.isReady) return;
@@ -40,8 +42,7 @@ function Post(props) {
   }
 
   const deletePost = () => {
-    if (window.confirm("Are you sure you want to delete post?"))
-      props.deletePost(id);
+    if (window.confirm(t("common:Message.Confirmation"))) props.deletePost(id);
     router.push("/");
   };
 
@@ -61,10 +62,10 @@ function Post(props) {
         {props.auth?.user?.role === "admin" ||
         post?.creator === props.auth?.user?.id ? (
           <div className="col d-flex justify-content-between">
-            <h5>If information was outdated, please update it or delete it </h5>
+            <h5>{t("common:PostId.Info")} </h5>
             <div>
               <a className="btn btn-text text-primary" onClick={deletePost}>
-                DELETE
+                {t("common:ChangeButton.DeleteButton")}
               </a>
               <a
                 className="btn btn-text text-primary"
@@ -72,7 +73,7 @@ function Post(props) {
                 data-bs-target="#exampleModal"
                 onClick={handleShow}
               >
-                EDIT
+                {t("common:ChangeButton.EditButton")}
               </a>
             </div>
           </div>
@@ -82,7 +83,7 @@ function Post(props) {
         <div className="row justify-content-center">
           <div className="col-11">
             <div className="comment">
-              <h6>Comments</h6>
+              <h6> {t("common:PostId.Comments")}</h6>
 
               {/* Comment */}
               <form onSubmit={handleSubmit(SubmitForm)}>
@@ -109,11 +110,11 @@ function Post(props) {
                         type="submit"
                         className="btn btn-primary btn-block"
                       >
-                        Publish
+                        {t("common:ChangeButton.PublishButton")}
                       </button>
                     ) : (
                       <button className="btn btn-primary btn-block" disabled>
-                        Publish
+                        {t("common:ChangeButton.PublishButton")}
                       </button>
                     )}
                   </div>
